Improved structure of LogMessage and LogSegment so they can be default constructed, or constructed with formatted input.
Also made a new method in LogMessage that will be used by the AsyncLogger that allows moving the calling object into another.

Since there are no more unique_ptr segments, made a method in LogSegment that allows seeing if the object has data (its text buffer is not empty).

Finally, made LogMessage append the prefix it has into the message body if a new prefix is set and there was one already.